third_party/WebKit/LayoutTests/platform/win/external/wpt/IndexedDB/idbrequest_error-expected.txt

This is a testharness.js-based test.
PASS IDBRequest.error throws if ready state is pending 
Harness: the test ran to completion.
